Removed eclipse warning only
git-svn-id: svn://www.xmlblaster.org/xmlBlaster/trunk@13936 00fa6119-6a0f-0410-b6da-8b96a2b62773

+++ xmlBlaster/src/java/org/xmlBlaster/contrib/dbwatcher/convert/Base64.java
@@ -65,9 +65,9 @@ public class Base64 {
         int n = 0;
 
         while (i <= end) {
-            int d = ((((int) data [i]) & 0x0ff) << 16) 
-                | ((((int) data [i+1]) & 0x0ff) << 8)
-                | (((int) data [i+2]) & 0x0ff);
+            int d = ((( data [i]) & 0x0ff) << 16) 
+                | ((( data [i+1]) & 0x0ff) << 8)
+                | (( data [i+2]) & 0x0ff);
 
             buf.append (charTab [(d >> 18) & 63]);
             buf.append (charTab [(d >> 12) & 63]);
@@ -84,8 +84,8 @@ public class Base64 {
 
 
         if (i == start + len - 2) {
-            int d = ((((int) data [i]) & 0x0ff) << 16) 
-                | ((((int) data [i+1]) & 255) << 8);
+            int d = ((( data [i]) & 0x0ff) << 16) 
+                | ((( data [i+1]) & 255) << 8);
 
             buf.append (charTab [(d >> 18) & 63]);
             buf.append (charTab [(d >> 12) & 63]);
@@ -93,7 +93,7 @@ public class Base64 {
             buf.append ("=");
         }
         else if (i == start + len - 1) {
-            int d = (((int) data [i]) & 0x0ff) << 16;
+            int d = (( data [i]) & 0x0ff) << 16;
 
             buf.append (charTab [(d >> 18) & 63]);
             buf.append (charTab [(d >> 12) & 63]);
@@ -106,11 +106,11 @@ public class Base64 {
 
     static int decode (char c) {
         if (c >= 'A' && c <= 'Z') 
-            return ((int) c) - 65;
+            return ( c) - 65;
         else if (c >= 'a' && c <= 'z') 
-            return ((int) c) - 97 + 26;
+            return ( c) - 97 + 26;
         else if (c >= '0' && c <= '9')
-            return ((int) c) - 48 + 26 + 26;
+            return (c) - 48 + 26 + 26;
         else switch (c) {
         case '+': return 62;
         case '/': return 63;
